+++ Tests/Resources/app/AppKernel.php
@@ -32,8 +32,15 @@ class AppKernel extends TestKernel
             new \Symfony\Cmf\Bundle\SeoBundle\CmfSeoBundle(),
             new \Burgov\Bundle\KeyValueFormBundle\BurgovKeyValueFormBundle(),
             new \Symfony\Cmf\Bundle\RoutingBundle\CmfRoutingBundle(),
-            new Symfony\Cmf\Bundle\CoreBundle\CmfCoreBundle(),
+            new \Symfony\Cmf\Bundle\CoreBundle\CmfCoreBundle(),
         ));
+
+        if (class_exists('Symfony\Cmf\Bundle\ResourceRestBundle\CmfResourceRestBundle')) {
+            $this->addBundles(array(
+                new \Symfony\Cmf\Bundle\ResourceBundle\CmfResourceBundle(),
+                new \Symfony\Cmf\Bundle\ResourceRestBundle\CmfResourceRestBundle(),
+            ));
+        }
     }
 
     public function registerContainerConfiguration(LoaderInterface $loader)